VotingMachine.py
This program is a voting machine where the user can input candidates and then cast a vote on the candidate in order to determine the winner. The program is not perfect, but it did teach me how to use nested loops more responsibly. In this case, the end of the program resulted in a buggy error which I will fix soon.